合同优化

+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/response/applycontract/ApplyContractResVo.java
@@ -43,7 +43,7 @@ public class ApplyContractResVo {
     private Integer paymentPeriod;
 
     @ApiModelProperty("配送费")
-    private Long shippingFee;
+    private BigDecimal shippingFee;
 
     @ApiModelProperty("送货费承担方(甲方,乙方承担)")
     private Byte deliveryCharges;
