+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/response/contract/ContractResVo.java
@@ -48,7 +48,7 @@ public class ContractResVo {
     private Integer paymentPeriod;
 
     @ApiModelProperty("配送费")
-    private Long shippingFee;
+    private BigDecimal shippingFee;
 
     @ApiModelProperty("送货费承担方(甲方,乙方承担)")
     private Byte deliveryCharges;
